Crate vfs_node

This crate contains a very basic, generic concrete implementation of the Directory and File traits. The VFSDirectory and VFSFile are intended to be used as regular nodes within the filesystem that require no special functionality as well as for inspiration for creating other concrete implementations s of the Directory and File traits.

Structs

VFSDirectory

A struct that represents a node in the VFS