Struct vfs_node::VFSDirectory

pub struct VFSDirectory {
    pub name: String,
    pub children: BTreeMap<String, FileOrDir>,
    pub parent: WeakDirRef,
}

A struct that represents a node in the VFS

Fields
name: String

The name of the directory

children: BTreeMap<String, FileOrDir>

A list of child filesystem nodes

parent: WeakDirRef

A weak reference to the parent directory

Implementations

impl VFSDirectory

pub fn new(name: String, parent: &DirRef) -> Result<DirRef, &'static str>

Creates a new directory and passes a pointer to the new directory created as output